Add auto retry for package manager based commands
Case RE-996: This makes it so that we automatically retry package
manager commands that fail on the first attempt.  The reason for this
change is that there is a network element involved that can sometimes
fail due to temporary mirror problems and the command can pass after
waiting for a few seconds.

Changelog: Add auto retry for package manager based commands