Update strings to be language agnostic for Ubuntu
Case RE-945:

Audited all the strings for RHEL specific language &
updated to be inclusive for Ubuntu. (However, strings
output by code that only runs on RHEL systems were not
updated.) Also, the --help output was not updated as that
is to be addressed in another case.

Changelog: Update strings to be language agnostic for Ubuntu