Un-invert logic of upgrade_distro_manually
During resolution of RE-670, it was discovered that PR #554, which
changed `should_run_distro_upgrade` to `upgrade_distro_manually`, failed
to account for the fact that this change of name inverts the sense of
the logic. Since none of the users of this method embed it in a
logically complicated way, and since there aren't too many users, modify
the method and its users to restore correct semantics.