Do not restore config files for dropped packages
Case RE-1179: Previously, we did not filter out dropped packages from
the list of EA4 packages that had config files that needed restored.
This could result in odd bugs where a file was put in place that should
not have been.

Changelog: Do not restore config files for dropped packages