Force mysqlcheck to use UTF-8
Case RE-1262: The `-A` option for the `mysqlcheck` utility provided by
MySQL/MariaDB gets the full list of databases/tables using the `SHOW
DATABASES/TABLES` commands.  Names of databases, tables, columns, and
other objects are permitted to contain any Unicode codepoint, but they
are documented to be stored as UTF-8, meaning that it should be safe to
assume that there won't be non-ASCII-compatible single-byte encodings in
use. Our use of the `C` locale in ELevate turns off UTF-8 support and
thus breaks `mysqlcheck`, because the environment variables we use to
enforce this cause `mysqlcheck` to also use this locale for
client-server communication, meaning that when a name of something has
Unicode characters, the server can't return it properly as part of `SHOW
DATABASES/TABLES`, and `mysqlcheck` thus can't direct the server to
check it.

The easiest way to get `mysqlcheck` to use UTF-8 is to use the
`--default-character-set=` flag. Pass the `utf8mb4` charset as the
argument for this option.

Changelog: Default MySQL/MariaDB database checks to a UTF-8 character
 set.