Ignore warnings about legacy trusted.gpg keyring
Case RE-1356: APT warns when a deprecated and less-secure
but traditional mechanism is used to configure GPG keys to verify
package signatures. This upsets the ELevate process and causes it to
bail out. Since we're not here to apply commentary on the way package
signatures are verified, ignore this warning for the purposes of error
checking.

Changelog: Ignore warnings on Ubuntu about deprecated storage of GPG keys.