Satisfy 8-to-9 dependence on VDO
Case RE-1387: VDO is a volume type that was originally proprietary but
was bought and open-sourced by Red Hat. Red Hat forces VDO users who
want to upgrade from RHEL 8 to 9 using leapp to convert their volumes
so that management no longer happens with the custom VDO tools and
instead uses LVM, which can now support these volumes. This requires a
conversion step which Red Hat was *not* comfortable having to occur
during the Leapp upgrade itself, so Leapp inhibits if it cannot confirm
that VDO is not in use, something for which it needs the VDO tools to be
installed.

There is an option to skip this check, but it seems easy enough to
install the VDO tools whenever we install Leapp on AlmaLinux 8, etc.
Therefore, create an Elevate::OS property that only returns true for
upgrades from 8 to 9, and install the `vdo` package when that is true,
plus a couple other conditions. Since it seems exceptionally unlikely
that any of our customers actually use VDO without having the tools
installed already, we should not have to worry about blocking during
`--check`.

Changelog: Install VDO tools to satisfy a Leapp dependency on EL8-based
 systems.