Ensure support of SHA1-signed packages
Case RE-1388: We and our software partners have a problem: AlmaLinux 9
does not support SHA-1 by default, but many of our software packages,
including those published for AL9, still build with SHA-1 signatures.
Leapp detects the AL8 versions of these packages and inhibits upgrade
unless the system cryptographic policy is set to explicitly allow SHA-1.
Unfortunately, AL8 does not ship with a sub-policy to explicitly enable
SHA-1, because AL8 supports SHA-1 in the default configuration.

For this reason, check that a supported cryptographic policy is enabled,
and if the default is in use, inject a definition for the SHA-1
sub-policy into the system and instruct the system to use it before
upgrade, and clean up after upgrade.

Changelog: Support systems with SHA1-signed packages.