Run leapp upgrade with UTF-8 locale
Case RE-1386: During creation of the initrd used by Leapp to perform the
OS upgrade, systemctl spits out UTF-8. This is a problem, because the
Python logging stream is configured to accept ASCII alone, even though
`leapp upgrade` alters its own environment to use a UTF-8 locale. Python
is probably doing this because it sees that we started Leapp using the C
locale.

In order to work around this issue until Leapp is changed upstream to
fix this issue, run `leapp upgrade` with the `en_US.UTF-8` locale.

Changelog: Work around encoding bug in Leapp.