Execute WPToolkit component sooner during post distro upgrade stage
Case RE-1395: Elevate is failing on u20->u22 upgrades while running
sysup after the distro upgrade is completed.  This is due to GPG issues
related to wp-toolkit.  This is a known issue and can be resolved by
reinstalling wp-toolkit using the provided installer:

https://support.plesk.com/hc/en-us/articles/12377139806359-Plesk-update-fails-GPG-error-The-following-signatures-couldn-t-be-verified-because-the-public-key-is-not-available-NO-PUBKEY

We already do that as part of the WPToolkit component during
post_distro_upgrade so this change updates the ordering of this
component so that it runs before sysup in order to avoid the GPG key
issue.

Changelog: Change ordering of components so that the WPToolkit component
 runs sooner during the post distro upgrade stage.