Add blocker for when ifcfg files are missing the TYPE parameter
Case RE-1535: Leapp has a check for A8->A9 upgrades to ensure that the
server will successfully convert from using network-scripts to
NetworkManager.  This change adds a blocker to elevate, so that we
natively block in a manner similar to the leapp check and provide a
better error message to customers to make it easier for them to overcome
this issue.

Changelog: Add blocker for when ifcfg files are missing the TYPE
 parameter on upgrades from A8->A9