Account for inherited domains when preserving PHP version
Case RE-1573: Previously, this logic blindly set the domain to use the
same PHP version that it used prior to upgrading the OS without
accounting for inherited domains.  This causes problems on CloudLinux
upgrades where the site is using PHP selector since PHP selector
requires that a domain be inherited before it can take over managing PHP
for the domain.

Changelog: Account for inherited domains when preserving PHP version