Increase timout in systemd oneshot file from 90s to 15min
Case RE-1602: fix-cpanel-perl runs as a ExecStartPre script.  This can
easily reach the default timeout of 90s on slower hardware or on servers
with slower network connections.  As such, we are going to increase this
timeout to 15min to give the script plenty of time to complete.

Changelog: Increate timeout in systemd unit file from default of 90s to
 15min