Increase timeout for verify_upgraded_os step
Case RE-1602:  The command timeout was set to 1 minute which led to the
integration test sometimes timing out.  It is weird that the integration
could ever take this long but considering how long it takes for elevate
to run, lets just bump this timeout as it not like this CI is something
that is ever going to be snappy anyway and having to start over at this
stage due to a timeout is very frustrating while providing no value.

Changelog: