Allow cPanel version to be newer than tier
Case RE-1604: Currently, when ELevate does the cPanel version check, it
has to match whatever the major version tier is exactly. However, it can
be the case that a named tier is on a newer build than the version tier,
which causes the existing check to fail when it should not. Change the
equality check to allow the version to be greater. (The function
checking this already assumes that both versions being compared have the
same major version, though this goes awry when the existing version is a
dev release. This case is caught, however.)

Changelog: Do not block the upgrade when the existing cPanel version is
  newer than what is stated in the major version's tier.