Move logic to add blocker for invalid syntax
Case RE-1647: Previously, the blocker invalid yum syntax would get added
prior to the autofix logic executing which resulted in servers getting
blocked in start mode even though they resolved the issue.  This meant
that executing the script with --start a second time would not get
blocked without any user interaction.  This is confusing for the user
and results in a CI failure now that the ELS repo has invalid syntax for
AlmaLinux 8 machines.

Changelog: Attempt to fix invalid yum repo syntax before blocking on it
 in start mode